requirements.txt
Flask==2.2.5           # Core web framework
Flask-SQLAlchemy==3.0.2 # ORM
Flask-Login==0.6.2     # User authentication
psycopg2-binary==2.9.5 # Postgres driver for Render
Flask-Mail==0.9.1      # Email support
Flask-Migrate==4.0.4   # DB migrations (required for flask db upgrade)
Flask-Babel==3.1.0     # i18n/l10n
Flask-Compress==1.13   # Gzip/Brotli compression
Flask-Caching==2.0.2   # Caching
Flask-SocketIO==5.3.4  # Real-time features (optional)
Werkzeug==2.2.3        # WSGI utils
SQLAlchemy==1.4.41     # ORM core
stripe==5.4.0          # Stripe payments
redis==4.5.4           # Redis (optional, for Celery/caching)
celery==5.2.7          # Celery (optional, for background tasks)
requests==2.28.2       # HTTP requests
Pillow==9.5.0          # Image processing
qrcode==7.4.2          # QR code generation
pyotp==2.8.0           # 2FA support
flask-cors
cryptography==41.0.1   # Security
# scikit-learn==1.2.2  # ML (optional)
# pandas==1.5.3        # Data analysis (optional)
# numpy==1.24.2        # Data analysis (optional)
# opencv-python==4.7.0.72 # Image processing (optional)
python-socketio==5.7.2 # Real-time (optional)
eventlet==0.33.3       # Async support (optional)
gunicorn==20.1.0       # Production WSGI server (required for Render)
python-dotenv==1.0.0   # .env file support (optional)
Jinja2==3.1.2          # Templating
MarkupSafe==2.1.2      # Templating
itsdangerous==2.1.2    # Security
click==8.1.3           # CLI utils
blinker==1.6.2         # Signals

psycopg2-binary
psycopg2-binary
